initial extraction of logic from backup controller
