+++ pkg/controller/restore_controller.go
@@ -621,3 +621,12 @@ func (l *restoreLogger) done(log logrus.FieldLogger) (io.Reader, error) {
 func (l *restoreLogger) closeAndRemove(log logrus.FieldLogger) {
 	closeAndRemoveFile(l.file, log)
 }
+
+func closeAndRemoveFile(file *os.File, log logrus.FieldLogger) {
+	if err := file.Close(); err != nil {
+		log.WithError(err).WithField("file", file.Name()).Error("error closing file")
+	}
+	if err := os.Remove(file.Name()); err != nil {
+		log.WithError(err).WithField("file", file.Name()).Error("error removing file")
+	}
+}
